reorganized the route structure, added a get and delete route for users
