+++ routes/user_functions/test.js
@@ -1,0 +1,5 @@
+const test = (req, res) => {
+  res.send('drunk with POWER');
+};
+
+module.exports = test;
